final touch, fix comments, bump version number to force migration

+++ database/migrations/2020_12_12_203153_migrate_admin_user.php
@@ -31,7 +31,9 @@ class MigrateAdminUser extends Migration
 	 */
 	public function down()
 	{
-		$user = User::find(0);
-		$user->delete();
+		if (Schema::hasTable('users')) {
+			$user = User::find(0);
+			$user->delete();
+		}
 	}
 }
